pyproject.toml
[project]
name = "price-tracker-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "django==4.2.20",
    "django-celery-beat>=2.7.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "environs>=14.1.1",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.1",
    "requests>=2.32.3",
]
